Fix year-2038-Problem
Without this change, timestamps after 2038-01-19 cannot be represented
in a signed 32-bit int.

I'm changing it to int64_t instead of time_t because of the reason
given when the `TimeValue` type was initially added in

commit 21288cb4c7da5a44f54e8e0a292b1026cdbcaeb5

    Author: Dave Mielke <dave mielke.cc>
    Date:   Tue Apr 3 06:00:00 2012 +0000

    timing.h should present the time in a platform-independent way. (dm)

    git-svn-id: svn://mielke.cc/main/brltty@6202 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/timing.c
@@ -213,7 +213,7 @@ expandTimeValue (const TimeValue *value, TimeComponents *components) {
 }
 
 size_t
-formatSeconds (char *buffer, size_t size, const char *format, int32_t seconds) {
+formatSeconds (char *buffer, size_t size, const char *format, int64_t seconds) {
   time_t time = seconds;
   struct tm description;
 
@@ -283,7 +283,7 @@ millisecondsTillNextSecond (const TimeValue *reference) {
 long int
 millisecondsTillNextMinute (const TimeValue *reference) {
   TimeValue time = *reference;
-  int32_t *seconds = &time.seconds;
+  int64_t *seconds = &time.seconds;
 
   time.nanoseconds = 0;
   *seconds /= SECS_PER_MIN;
